+++ src/main/java/com/seguimiento/repository/usuario/UsuarioRepository.java
@@ -5,6 +5,6 @@ import com.seguimiento.model.Usuario;
 
 public interface UsuarioRepository extends JpaRepository<Usuario, Long>{
 	
-	Usuario findByDni(int dni);
+	Usuario findByDni(long dni);
 	
 }
